backend/requirements.txt
# FastAPI Backend Requirements
# Python 3.11+

# Core Framework
fastapi==0.129.0
uvicorn[standard]==0.40.0
pydantic==2.12.5
pydantic-settings==2.12.0

# Database
sqlalchemy==2.0.46
pymysql==1.1.2
cryptography==46.0.5
alembic==1.18.4

# Authentication
python-jose[cryptography]==3.5.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.22

# CORS & Security
python-dotenv==1.2.1
slowapi==0.1.9

# CSV / Excel Processing
openpyxl==3.1.5

# PDF Generation
reportlab==4.4.10

# Email & Reports
sendgrid==6.12.5
APScheduler==3.11.2

# System Monitoring
psutil==7.2.2

# QR Code Generation
qrcode[pil]==8.2
simpy==4.1.1
